[FIX] reorganized stripe controllers in a single file, made export for stripe var
