+++ src/utils/stripeClient.mjs
@@ -1,0 +1,6 @@
+import Stripe from 'stripe';
+
+const stripeKey = process.env.STRIPE_SECRET_KEY;
+const stripe = new Stripe(stripeKey);
+
+export default stripe;
